[DOC] Started commit
Due to an issue with bad data management on a notebook, I decide to restart the project from the beginning, so this is the main code needed to start as a project.

For project_path.py at notebooks dir, you will need to adapt or copy it according to the levels that you move it.